Add dirs to PATH, wrap for-loops (and vars) in fn
In adding an additional dir to my PATH, I hit a bug where my for-loop
counter would get changed by a child script that was using the same
variable for its loop counter, thus reducing the value of the counter
and causing an infinite loop. This was fixed by wrapping all loops in
their own functions and making the variables locally scoped.