Add script to increase max mouse sensitivity.
Useful on some machines (usually Debian machines, I've found) that don't
have a high enough max sensitivity for the mouse.

I usually set this to run on startup, for whatever the default pointer
device is. In this case, it's my Logitech g305, which is always plugged
in to my workstation.