fix: local-ize local variable, disable pyenv init
When installing pyenv on my local machine, it activated my pyenv rc
file, which called path_prepend... this made loading my .bashrc hang. I
eventually found out it was because the path_prepend function was using
a loop-counter variable "i" without using a local copy of it, resulting
it it resetting an "i" counter variable in a higher scope, causing an
infinite loop.